Separated ExposureDataRepository from UserDataRepository.
