+++ Covid19Radar/Covid19Radar/ViewModels/Settings/DebugPageViewModel.cs
@@ -19,6 +19,7 @@ namespace Covid19Radar.ViewModels
         private readonly ITermsUpdateService _termsUpdateService;
         private readonly IExposureConfigurationRepository _exposureConfigurationRepository;
         private readonly IUserDataRepository _userDataRepository;
+        private readonly IExposureDataRepository _exposureDataRepository;
         private readonly AbsExposureNotificationApiService _exposureNotificationApiService;
         private readonly AbsExposureDetectionBackgroundService _exposureDetectionBackgroundService;
         private readonly ICloseApplicationService _closeApplicationService;
@@ -85,8 +86,8 @@ namespace Covid19Radar.ViewModels
 
             var exposureNotificationStatus = await _exposureNotificationApiService.IsEnabledAsync();
 
-            var dailySummaryCount = (await _userDataRepository.GetDailySummariesAsync(AppConstants.DaysOfExposureInformationToDisplay)).Count();
-            var exposureWindowCount = (await _userDataRepository.GetExposureWindowsAsync(AppConstants.DaysOfExposureInformationToDisplay)).Count();
+            var dailySummaryCount = (await _exposureDataRepository.GetDailySummariesAsync(AppConstants.DaysOfExposureInformationToDisplay)).Count();
+            var exposureWindowCount = (await _exposureDataRepository.GetExposureWindowsAsync(AppConstants.DaysOfExposureInformationToDisplay)).Count();
 
             // ../../settings.json
             var settings = new[] {
@@ -99,7 +100,7 @@ namespace Covid19Radar.ViewModels
                 $"PrivacyPolicyUpdatedDateTimeUtc: {privacyPolicyUpdateDateTimeUtc}",
                 $"StartDate: {_userDataRepository.GetStartDate().ToLocalTime().ToString("F")}",
                 $"DaysOfUse: {_userDataRepository.GetDaysOfUse()}",
-                $"Legacy-V1 ExposureCount: {_userDataRepository.GetV1ExposureCount(AppConstants.DaysOfExposureInformationToDisplay)}",
+                $"Legacy-V1 ExposureCount: {_exposureDataRepository.GetV1ExposureCount(AppConstants.DaysOfExposureInformationToDisplay)}",
                 $"DailySummaryCount: {dailySummaryCount}",
                 $"ExposureWindowCount: {exposureWindowCount}",
                 $"LastProcessTekTimestamp: {lastProcessTekTimestampsStr}",
@@ -133,6 +134,7 @@ namespace Covid19Radar.ViewModels
             ITermsUpdateService termsUpdateService,
             IExposureConfigurationRepository exposureConfigurationRepository,
             IUserDataRepository userDataRepository,
+            IExposureDataRepository exposureDataRepository,
             AbsExposureNotificationApiService exposureNotificationApiService,
             AbsExposureDetectionBackgroundService exposureDetectionBackgroundService,
             ICloseApplicationService closeApplicationService,
@@ -143,6 +145,7 @@ namespace Covid19Radar.ViewModels
             _termsUpdateService = termsUpdateService;
             _exposureConfigurationRepository = exposureConfigurationRepository;
             _userDataRepository = userDataRepository;
+            _exposureDataRepository = exposureDataRepository;
             _exposureNotificationApiService = exposureNotificationApiService;
             _exposureDetectionBackgroundService = exposureDetectionBackgroundService;
             _closeApplicationService = closeApplicationService;
@@ -199,7 +202,7 @@ namespace Covid19Radar.ViewModels
 
         public Command OnClickRemoveExposureInformation => new Command(async () =>
         {
-            _userDataRepository.RemoveExposureInformation();
+            _exposureDataRepository.RemoveExposureInformation();
             await UpdateInfo("RemoveExposureInformation");
         });
